Check hardware: - CTD ok - FOG ok - GPS-AHRS ok - DVL, PA500, Echo da verificare misure - USBL-Pinger da verificare pos e dati

+++ common/data/Filter_Ang_status.h
@@ -19,7 +19,7 @@ struct Filter_Ang_tlm_packet
 	int64 r_r;
 	int64 Kr;
 	int64 br;
-	int32 execution;
+	int64 execution;
 
 	void zero()
 	{
